вынес типы для отображения полей на русском
